Moved portfolio class to singleton Began adding import of data from other portfolios

+++ MyPersonalIndex/classes/account.cpp
@@ -23,5 +23,8 @@ void account::save(const int &portfolioID)
 
 void account::remove() const
 {
+    if (this->id == -1)
+        return;
+
     queries::deleteItem(queries::table_Acct, this->id);
 }
